chore: moved Wizard to ui-client
