--- apps/meteor/client/components/Wizard/WizardActions.tsx
@@ -1,16 +1,0 @@
-import { Box, ButtonGroup } from '@rocket.chat/fuselage';
-import type { ReactNode } from 'react';
-
-type WizardActionsProps = {
-	children: ReactNode;
-};
-
-const WizardActions = ({ children }: WizardActionsProps) => {
-	return (
-		<Box className='steps-wizard-footer' pbs={24} display='flex' justifyContent='end'>
-			<ButtonGroup>{children}</ButtonGroup>
-		</Box>
-	);
-};
-
-export default WizardActions;
