+++ packages/ui-client/src/components/Wizard/index.ts
@@ -1,8 +1,14 @@
 export { default } from './Wizard';
+export { default as Wizard } from './Wizard';
 export { default as WizardContent } from './WizardContent';
 export { default as WizardTabs } from './WizardTabs';
 export { default as WizardActions } from './WizardActions';
 export { default as WizardBackButton } from './WizardBackButton';
 export { default as WizardNextButton } from './WizardNextButton';
+export { default as StepsLinkedList } from './lib/StepsLinkedList';
+export { default as StepNode } from './lib/StepNode';
+export * from './lib/StepNode';
+export * from './WizardContext';
 export * from './useWizardContext';
 export * from './useWizard';
+export * from './mocks/createMockWizardApi';
