+++ packages/ui-client/src/components/index.ts
@@ -15,3 +15,4 @@ export { default as AnnouncementBanner } from './AnnouncementBanner';
 export { default as UserAutoComplete } from './UserAutoComplete';
 export * from './GenericMenu';
 export * from './Modal';
+export * from './Wizard';
